+++ src/settings/LoanPolicy/LoanPolicyDetail.js
@@ -2,7 +2,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {
   injectIntl,
-  intlShape,
   FormattedMessage,
 } from 'react-intl';
 import {
@@ -32,7 +31,7 @@ import { intervalPeriods } from '../../constants';
 
 class LoanPolicyDetail extends React.Component {
   static propTypes = {
-    intl: intlShape.isRequired,
+    intl: PropTypes.object,
     initialValues: PropTypes.object,
     stripes: stripesShape.isRequired,
     parentResources: PropTypes.shape({